feat: fix loading mails bug, implemented 'sync' and ' delete' buttons

+++ database.sql
@@ -27,7 +27,7 @@ CREATE TABLE IF NOT EXISTS public.tags (
 CREATE TABLE IF NOT EXISTS public.mails (
                                             id          SERIAL PRIMARY KEY,
                                             subject     VARCHAR(255) NOT NULL,
-    mail_id     INTEGER NOT NULL,
+    mail_id     BIGINT NOT NULL,
     text        TEXT NOT NULL,
     received_at TIMESTAMP,
     is_read     BOOLEAN DEFAULT FALSE,
